requirements.txt
flask
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
yt-dlp
librosa
numpy
Pillow
